fix token page speed
